Decomposing top level App render flow
