+++ components/dashboard/src/app/AppLoading.tsx
@@ -1,0 +1,12 @@
+/**
+ * Copyright (c) 2022 Gitpod GmbH. All rights reserved.
+ * Licensed under the GNU Affero General Public License (AGPL).
+ * See License.AGPL.txt in the project root for license information.
+ */
+
+import { FunctionComponent } from "react";
+
+// TODO: Would be great to show a loading UI if it's been loading for awhile
+export const AppLoading: FunctionComponent = () => {
+    return <></>;
+};
